@@ -2,9 +2,7 @@
 // SPDX-License-Identifier: MIT-0
 
 const awsPath = '/Prod';
-export const rootPath: string = window.location.href.includes(awsPath)
-  ? `${awsPath}/`
-  : '/';
+export const rootPath: string = window.location.href.includes(awsPath) ? `${awsPath}/` : '/';
 
 const routes = {
   HOME: `${rootPath}`,
