@@ -7,7 +7,7 @@ import MeetingFormSelector from '../../containers/MeetingFormSelector';
 import { StyledLayout } from './Styled';
 import { VersionLabel } from '../../utils/VersionLabel';
 
-const Home = () => (
+const Home: React.FC = () => (
   <StyledLayout>
     <MeetingFormSelector />
     <VersionLabel />
